% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scores.R
\name{create_scores}
\alias{create_scores}
\title{Create scores}
\usage{
create_scores(
  resp,
  vars,
  select,
  scoring = NULL,
  score_name = "score",
  num_cat = "num_cat",
  xsi_fixed = NULL,
  rotation = NULL,
  valid = NULL,
  mvs = NULL,
  missing_by_design = -54,
  wle = TRUE,
  sum_score = FALSE,
  sum_select = NULL,
  metap = FALSE,
  meta_variable = NULL,
  meta_score_name = NULL,
  meta_select = NULL,
  control_tam = NULL,
  control_wle = NULL,
  pweights = NULL,
  poly2dich = TRUE,
  save = TRUE,
  return = FALSE,
  name_group = NULL,
  overwrite = FALSE,
  path_results = "Results",
  path_table = "Tables",
  warn = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{score_name}{character; name of the scores -- WITHOUT extension (e.g.,
reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{num_cat}{string; defines name of numeric variable in vars that indicates
the maximum number of response options for the items}

\item{xsi_fixed}{named numerical vector; contains fixed item difficulties as
elements and item names as names of elements}

\item{rotation}{character vector; contains the variable name indicating the
test rotation}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{missing_by_design}{numeric; user defined missing value for missing by
design (is necessary for calculating N_administered)}

\item{wle}{logical; whether to estimate WLEs}

\item{sum_score}{logical; whether to calculate sum scores}

\item{sum_select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis
(if identical to select, argument can be empty)}

\item{metap}{logical; whether to calculate metacognition score}

\item{meta_variable}{string; defines name of meta competence variable in resp}

\item{meta_score_name}{string; name of the meta competence scores -- WITHOUT
extension (e.g., reg4 instead of reg4_sc1 or mag12 instead of mag12_sc1u)}

\item{meta_select}{string; defines name of logical variable in vars that
indicates which items to use for the meta cognition analysis
(if identical to select, argument can be empty)}

\item{control_tam}{list; control argument as passed to tam.mml.mfr()}

\item{control_wle}{list; can contain Msteps and/or convM as to pass to tam.wle()
as elements of the list}

\item{pweights}{numeric vector; person weights for current measurement point
passed to TAM-functions}

\item{poly2dich}{logical; indicates whether count only correctly scored
binary or FULLY correctly scored PC items}

\item{save}{logical; whether results shall be saved to hard drive}

\item{return}{logical; whether results shall be returned}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path_results}{string; defines path to folder where results shall be
saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{warn}{logical; whether to print warnings}
}
\description{
Create scores
}
