% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_analysis}
\alias{dif_analysis}
\title{DIF analyses - all in one function}
\usage{
dif_analysis(
  resp,
  vars,
  select,
  dif_vars,
  valid = NULL,
  mvs = NULL,
  scoring = NULL,
  overwrite = FALSE,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  include_mv = 200,
  min_val = 50,
  control = NULL,
  pweights = NULL,
  path_results = "Results",
  path_table = "Tables",
  suf_item_names = FALSE,
  name_group = NULL,
  verbose = FALSE,
  warn = TRUE,
  dif_threshold = 0.5,
  digits = 3
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{character vector; defines name(s) of logical variable(s) in vars
that indicates which items to use for analysis; if some of the \code{dif_vars}
come with a different set of analysis items, this argument becomes a
vector of \code{length(dif_vars)} containing the respective selection
variables in vars}

\item{dif_vars}{character vector; contains the variable names to be tested
for DIF (e.g., "gender")}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{include_mv}{numeric; identifies threshold for which group size missing
values should be included in analysis as an extra group (defaults to 200)}

\item{min_val}{integer; minimum number of valid responses for each group on
each item for the item to be included in the analysis (defaults to 50)}

\item{control}{list; function argument as passed to TAM-functions}

\item{pweights}{character; defines name of numerical variable in resp that
contains person weights as passed to TAM-function}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{suf_item_names}{logical; whether to output SUF item names in the .xlsx file
for items with collapsed categories}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{dif_threshold}{numeric scalar; indicates absolute threshold of
problematic DIF (defaults to 0.5)}

\item{digits}{integer; number of decimals for rounding}
}
\value{
(if return = TRUE) a list of:
  models: list with DIF model results for all variables defined in 'dif_vars'
  summaries: list with a summary of DIF results for all variables defined in 'dif_vars'
  tr_tables: table for TR with summary of results for all variables defined in 'dif_vars'
}
\description{
Testing for differential item functioning for binary and polytomous data.
Main effects and DIF effects models are estimated.
}
