% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_analysis.R
\name{dim_analysis}
\alias{dim_analysis}
\title{Dimensionality analysis}
\usage{
dim_analysis(
  resp,
  vars,
  select,
  scoring = "scoring",
  dim = NULL,
  valid = NULL,
  mvs = NULL,
  irtmodel = "PCM2",
  maxiter = 10000,
  snodes = 5000,
  return = FALSE,
  save = TRUE,
  print = TRUE,
  path_results = "Results",
  path_table = "Tables",
  name_group = NULL,
  overwrite = TRUE,
  verbose = FALSE,
  digits = 3
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{string; defines name of logical variable in vars that indicates
which items to use for the analysis}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; defaults to
"scoring"}

\item{dim}{character vector. contains names of all dimension variables
that shall be analyzed (NOTE: the dimensions must be coded as integers from 1
to the number of dimensions!)}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{irtmodel}{string; "1PL" for Rasch, "2PL" for 2PL, "PCM2" for PCM and
"GPCM" for GPCM analysis}

\item{maxiter}{max iterations as passed to the TAM function}

\item{snodes}{snodes as passed to the TAM function}

\item{return}{logical; whether results shall be returned}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{verbose}{verbose as passed to the TAM function}

\item{digits}{integer; number of decimals for rounding}
}
\value{
(if return = TRUE) list of results:
 analysis: list with results for each dimensionality model
 summary: list with summary of results (variance-covariance matrices & model fit).
}
\description{
Dimensionality analysis
}
