% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_report_figures.r
\name{FigWrightMap}
\alias{FigWrightMap}
\alias{FigWright_Map}
\title{Crop plot with Wright map and include footnote}
\usage{
FigWrightMap(file, tbl = 5, crop = "800x1010+5+50")

FigWright_Map(file, tbl = 5, crop = "800x1010+5+50")
}
\arguments{
\item{file}{A file path for the image.}

\item{tbl}{The table number for the footnote.}

\item{crop}{A string indicating how to crop the image in the format
"<width>x<height>\{+-\}<xoffset>\{+-\}<yoffset>"
(see http://www.imagemagick.org/Magick++/Geometry.html)}
}
\value{
An image.
}
\description{
Crop plot with Wright map and include footnote
}
