% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{grouped_irt_analysis}
\alias{grouped_irt_analysis}
\title{IRT analyses for several groups}
\usage{
grouped_irt_analysis(
  resp,
  vars,
  select,
  valid = NULL,
  mvs = NULL,
  missing_by_design = -54,
  scoring = NULL,
  plots = FALSE,
  save = TRUE,
  print = TRUE,
  return = FALSE,
  overwrite = FALSE,
  path_plots = "Plots",
  path_table = "Tables",
  path_results = "Results",
  suf_item_names = FALSE,
  digits = 3,
  verbose = FALSE,
  warn = TRUE,
  xsi_fixed_1p = NULL,
  xsi_fixed_2p = NULL,
  pweights = NULL,
  control_tam = NULL,
  control_wle = NULL
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{select}{list; contains name of item selection variable of each group as
elements and group name as in resp and vars as element names
(e.g. easy = 'easy_final')}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{missing_by_design}{numeric; user defined missing value for missing by
design (is necessary for calculating N_administered)}

\item{scoring}{string; defines name of numerical variable in vars that
contains the scoring factor to be applied to loading matrix; can be NULL for
Rasch model}

\item{plots}{logical; whether plots shall be created and saved to hard drive}

\item{save}{logical; whether results shall be saved to hard drive}

\item{print}{logical; whether results shall be printed to console}

\item{return}{logical; whether results shall be returned}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{path_plots}{string; defines path to folder where plots shall be saved}

\item{path_table}{string; defines path to folder where tables shall be saved}

\item{path_results}{string; defines path to folder where results shall be saved}

\item{suf_item_names}{logical; whether to output SUF item names in the .xlsx file
for items with collapsed categories}

\item{digits}{integer; number of decimals for rounding}

\item{verbose}{logical; whether to print processing information to console}

\item{warn}{logical; whether to print warnings (should be set to TRUE to
avoid problems with the data structure)}

\item{xsi_fixed_1p}{list; may contain argument xsi_fixed_1p from irt_analysis()
for each group as elements and group names as names of elements; may be NULL
or contain xsi_fixed only for some groups}

\item{xsi_fixed_2p}{list; may contain argument xsi_fixed_2p from irt_analysis()
for each group as elements and group names as names of elements; may be NULL
or contain xsi_fixed only for some groups}

\item{pweights}{character; defines name of numerical variable in resp that
contains person weights as passed to TAM-function}

\item{control_tam}{list; may contain argument control_tam from irt_analysis()
for each group as elements and group names as names of elements; may be NULL
or contain xsi_fixed only for some groups}

\item{control_wle}{list; may contain argument control_wle from irt_analysis()
for each group as elements and group names as names of elements; may be NULL
or contain xsi_fixed only for some groups}
}
\value{
(if return = TRUE) list with a list for each group with all results
(as return object of irt_analysis()) as list elements.
}
\description{
IRT analyses for several groups
}
