% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_analyses.R
\name{irt_summary}
\alias{irt_summary}
\title{Summary of IRT analysis}
\usage{
irt_summary(
  resp,
  vars,
  valid = NULL,
  mvs = NULL,
  missing_by_design = -54,
  results,
  disc = NULL,
  save = TRUE,
  path = "Tables",
  suf_item_names = FALSE,
  name_group = NULL,
  digits = 3,
  overwrite = FALSE,
  warn = TRUE,
  test = TRUE
)
}
\arguments{
\item{resp}{data.frame; contains item responses with items as variables and
persons as rows; y in \{0, 1\} for binary data and y in \{0, 1, ... k-1\} for
polytomous responses with k categories; missing values (default -999 to -1)
are coded as NA internally; additionally includes ID_t as a person identifier
and all variables that are further defined in the function arguments
includes all variables that are further defined in the function arguments}

\item{vars}{data.frame; contains information about items with items as rows;
includes variable 'item' containing item names; additionally includes all
variables that are further defined in the function arguments}

\item{valid}{string; defines name of logical variable in resp that indicates
(in)valid cases}

\item{mvs}{named integer vector; contains user-defined missing values}

\item{missing_by_design}{numeric; user-defined missing value for missing by
design (is necessary for calculating N_administered)}

\item{results}{list; return object of irt_model(); one parameter model}

\item{disc}{list; return object of irt_model(); two parameter model}

\item{save}{logical; whether results shall be saved to hard drive}

\item{path}{string; defines path to folder where table shall be saved}

\item{suf_item_names}{logical; whether to output SUF item names in the .xlsx file
for items with collapsed categories}

\item{name_group}{string; defines name of group used in analysis (e.g. 'easy')}

\item{digits}{integer; number of decimals for rounding}

\item{overwrite}{logical; whether to overwrite existing file when saving table}

\item{warn}{logical; whether to print warnings (should be set to TRUE)}

\item{test}{logical; whether to test data structure (should be set to TRUE)}
}
\value{
a data.frame (for TR) containing the item name, N, percentage correct,
  item difficulty, SE, WMNSQ, t, rit, item discrimination, Q3.
}
\description{
Create table with results of IRT analysis.
}
